requirements.txt
transformers>=4.40.0
torch>=2.2.0
accelerate
python-dotenv
